requirements.txt
PyYAML
click
GitPython
docker
botocore
boto3
